Update PowerSteg.ps1
Swapped out modulo math with bitwise operator (AND) functionality to speed up execution further.  

Seeing mixed results, where it saves .5 seconds in some cases, and is about the same in others.  Seems like desteg is consistently faster, while ensteg can be a little faster or just the same as before.